Fixed delete

+++ server/routes/shopRouter.js
@@ -40,21 +40,24 @@ shopRouter.post(
             price,
             description,
           });
-          if (req.files.images) {
-            for (const file of req.files.images) {
-              const name = `${Date.now()}.webp`;
-              const outputBuffer = await sharp(file.buffer).webp().toBuffer();
-              await fs.writeFile(`./public/images/${name}`, outputBuffer);
-              await Image.create({ productId: newProduct.id, url: name, forConstructor: false });
+          try {
+            if (req?.files?.images) {
+              for (const file of req.files.images) {
+                const name = `${Date.now()}.webp`;
+                const outputBuffer = await sharp(file.buffer).webp().toBuffer();
+                await fs.writeFile(`./public/images/${name}`, outputBuffer);
+                await Image.create({ productId: newProduct.id, url: name, forConstructor: false });
+              }
             }
+            if (req?.files?.cover[0]) {
+              const name = `${Date.now()}.webp`;
+              const outputBuffer = await sharp(req.files.cover[0].buffer).webp().toBuffer();
+              await fs.writeFile(`./public/images/${name}`, outputBuffer);
+              await Image.create({ productId: newProduct.id, url: name, forConstructor: true });
+            }
+          } catch (err) {
+            return res.status(403).json({ message: 'Нет картинки' });
           }
-          if (req.files.cover[0]) {
-            const name = `${Date.now()}.webp`;
-            const outputBuffer = await sharp(req.files.cover[0].buffer).webp().toBuffer();
-            await fs.writeFile(`./public/images/${name}`, outputBuffer);
-            await Image.create({ productId: newProduct.id, url: name, forConstructor: true });
-          }
-          console.log(999999999999999999);
           const sizes = await Size.findAll();
           for (let i = 0; i < sizes.length; i++) {
             ProductSize.create({ productId: newProduct.id, sizeId: sizes[i].id, count: 50 });
@@ -117,16 +120,18 @@ shopRouter.put(
 
 shopRouter.delete('/products/:id', async (req, res) => {
   const user = await User.findOne({
-    where: { id: req.session.user.id },
+    where: { id: req?.session?.user?.id },
     include: { model: Role },
   });
   if (user.Role.id === 1) {
     const product = await Product.findByPk(req.params.id);
     const images = await Image.findAll({ where: { productId: product.id } });
+    const folder = await fs.readdir('./public/images/');
     for (const image of images) {
-      console.log(image);
-      await fs.unlink(`./public/images/${image.url}`);
-      await image.destroy();
+      if (folder.includes(image.url)) {
+        await fs.unlink(`./public/images/${image.url}`);
+      }
+        await image.destroy();
     }
     await ProductSize.destroy({ where: { productId: product.id } });
     product.destroy();
@@ -156,7 +161,6 @@ shopRouter.post('/orders', async (req, res) => {
   res.json(cant);
 });
 shopRouter.post('/cart/:productId', async (req, res) => {
-  console.log(req.params.productId);
   await Cart.create({
     userId: req.session.user.id,
     productSizeId: req.params.productId,
